+++ packages/@cmsonfire/cms-types/index.ts
@@ -1,0 +1,2 @@
+export * as CmsWidget from './lib/CmsonfireCmsWidget';
+export * as CmsCore from './lib/CmsonfireCmsCore';