+++ packages/@cmsonfire/cms-types/src/index.ts
@@ -1,2 +1,0 @@
-export * as CmsWidget from './CmsWidget';
-export * as CmsCore from './CmsCore';