+++ packages/@cmsonfire/cms-lib-template/CHANGELOG.md
@@ -1,9 +1,8 @@
-# @cmsonfire/cms-widget-text
+# @cmsonfire/cms-lib-template
 
-## 0.0.4
+## 0.0.2
 
 ### Patch Changes
 
-- fb3aad7: New build configuration (chore)
 - Updated dependencies [fb3aad7]
   - @cmsonfire/cms-types@0.0.4